lesson stats: don't commit no changes (#19)
There are times when there are [no changes to be had in the lesson stats](https://github.com/carpentries-incubator/carpentries-incubator.org/commit/fd9bdf236bd0523250f9168045b5b1cd421c94d0), for those times, we should not for a commit. This adds the `||` operator to pipe out to a valid command on error.